Fixes from lastest merge. Ignored some warnings. Fixed warnings related to casting to bool. Updated Windows code to use new otInit and otEnable APIs.

+++ src/core/openthread.cpp
@@ -942,7 +942,7 @@ ThreadError otSendIp6Datagram(otContext *aContext, otMessage aMessage)
         Ip6::Ip6::HandleDatagram(
             *static_cast<Message *>(aMessage),
             NULL,
-            (uint8_t)aContext->mThreadNetif.GetInterfaceId(),
+            aContext->mThreadNetif.GetInterfaceId(),
             NULL,
             true
         );
